Models

+++ app/lib/utils/zone.dart
@@ -1,7 +1,7 @@
 class Zone {
   static String getZone(double latitude, double longitude){
-    double lat = (latitude*100).ceilToDouble();
-    double lon = (longitude*100).floorToDouble();
+    double lat = (latitude*100).toInt().ceilToDouble();
+    double lon = (longitude*100).toInt().floorToDouble();
 
     return "${formatter(lat)};${formatter(lon)}";
   }
@@ -11,7 +11,6 @@ class Zone {
       number = number.abs();
       return "-${number.toStringAsFixed(0).padLeft(4, '0')}";
     }
-
     return "+${number.toStringAsFixed(0).padLeft(4, '0')}";
   }
 }
